+++ Bivrost/Classes/Coder/IntX.swift
@@ -1,0 +1,8 @@
+//
+//  IntX.swift
+//  Bivrost
+//
+//  Created by Luis Reisewitz on 26.09.17.
+//
+
+import BigInt
