+++ Bivrost/Classes/Coder/Solidity.swift
@@ -5,4 +5,5 @@
 //  Created by Luis Reisewitz on 26.09.17.
 //
 
+/// Blanket Holder for Solidity Types
 struct Solidity {}
