+++ Bivrost/Classes/Coder/UIntX.swift
@@ -1,5 +1,5 @@
 //
-//  UInt.swift
+//  UIntX.swift
 //  Bivrost
 //
 //  Created by Luis Reisewitz on 26.09.17.
