+++ Bivrost/Classes/Models/Contract.swift
@@ -7,8 +7,6 @@
 //
 //
 
-import Foundation
-
 /// Describes the interface of an Ethereum/Solidity Contract
 struct Contract {
     let elements: [Element]
