+++ Bivrost/Classes/Protocols/AbiValidating.swift
@@ -5,8 +5,6 @@
 //  Created by Luis Reisewitz on 19.09.17.
 //
 
-import Foundation
-
 protocol AbiValidating {
     /// Returns if the internal represented state is valid for usage in Ethereum ABI.
     var isValid: Bool { get }
